Implemented a better filtering mechanism for paths: - Added interfaces IUserList.cs, IUser.cs and IPasswdProvider.cs to the Interop namespace. - Implemented these interfaces in UserList.cs, User.cs and PasswdProvider.cs, respectively. - Added new class PathFilter.cs to the SurfaceArea namespace. Gave it a static UserList instance to allow redirection for automated testing without having to pass a value in on every construction. Implemented IXmlSerializable.cs so that this complex object structure can be preserved automagically in the process of OperatingParameters serialization/deserialization. Added enumerated type PathFilterType.cs. - Converted ExcludePaths and ExcludePathsWithComponents in OperatingParameters.cs into a list of PathFilters. Rewrote IsExcludedPath to use the new data. Updated BackupAgent.cs to ignore deletions in excluded paths and moves where both from & to are excluded paths in the monitor_PathMove and monitor_PathDelete handlers. Added unit tests of UserList.cs and PathFilter.cs to DeltaQ.RTB.Tests. Updated ConsolidateOldestBatch_should_omit_deleted_paths in RemoteFileStateCacheTests.cs to ensure that there are no collisions in Path values in generated fake file state objects.

+++ DeltaQ.RTB/Agent/BackupAgent.cs
@@ -107,7 +107,10 @@ namespace DeltaQ.RTB.Agent
 				}
 			}
 
-			AddActionToBackupQueue(new MoveAction(move.PathFrom, move.PathTo));
+			if (_parameters.IsExcludedPath(move.PathFrom) && _parameters.IsExcludedPath(move.PathTo))
+				VerboseDiagnosticOutput("Ignoring PathMove event on excluded paths: {0} => {1}", move.PathFrom, move.PathTo);
+			else
+				AddActionToBackupQueue(new MoveAction(move.PathFrom, move.PathTo));
 		}
 
 		void monitor_PathDelete(object? sender, PathDelete delete)
@@ -124,7 +127,10 @@ namespace DeltaQ.RTB.Agent
 				}
 			}
 
-			AddActionToBackupQueue(new DeleteAction(delete.Path));
+			if (_parameters.IsExcludedPath(delete.Path))
+				VerboseDiagnosticOutput("Ignoring PathDelete event on excluded path: {0}", delete.Path);
+			else
+				AddActionToBackupQueue(new DeleteAction(delete.Path));
 		}
 
 		public void PauseMonitor()
